% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{?}
\alias{?}
\title{Show help on Google Translate}
\usage{
`?`(e1, e2)
}
\arguments{
\item{e1, e2}{Arguments passed to \code{utils::`?` }}
}
\description{
The help document is translated to the language based on
\verb{getOption("transdoc_to, "en")}. See \code{language} for the available languages.
}
\examples{
\dontrun{
options(transdoc_to = "ja")
?base::c
}
}
