% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\docType{data}
\name{language}
\alias{language}
\title{Available language for translation}
\format{
A named character vector.
}
\usage{
language
}
\description{
The name or the value of this object can be specified to the \code{to} argument of
\code{translate()} or the \code{transdoc_to} argument of \code{option()}.
}
\keyword{datasets}
