% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print-methods}
\alias{print-methods}
\alias{print.transdoc_help}
\title{Print methods}
\usage{
\method{print}{transdoc_help}(x, to = "default", ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{to}{\code{doc} is translated to the language specified to \code{to}
(e.g., \code{"en"}, \code{"ja"}, \code{"de"}). The default value, \code{"default"},
is equivalent to \code{getOption("transdoc_to", "en")}.
For \verb{\%to\%}, the non-standard evaluation takes place for \code{to}.}

\item{...}{Arguments passed to the methods.}
}
\description{
Print methods for objects generated by the transdoc package.
Google Translate translates the input documents.
}
\examples{
\dontrun{
print(translate("This text will be translated to Japanese", "ja"))
print(translate("https://transdoc.atusy.net/", "ja"))
print(translate(?base::c, "ja"))
}

}
