% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\alias{\%to\%}
\alias{translate.character}
\alias{translate.help_files_with_topic}
\alias{translate.transdoc_help}
\title{Translate a text, web, or help document.}
\usage{
translate(doc, to = "default", from = "auto", ...)

doc \%to\% to

\method{translate}{character}(doc, to = "default", from = "auto", ...)

\method{translate}{help_files_with_topic}(doc, to = "default", from = "auto", ...)

\method{translate}{transdoc_help}(doc, to = "default", from = "auto", ...)
}
\arguments{
\item{doc}{A character vector of text, a string of URL, or an object returned
by \code{help} or \verb{?}.}

\item{to}{\code{doc} is translated to the language specified to \code{to}
(e.g., \code{"en"}, \code{"ja"}, \code{"de"}). The default value, \code{"default"},
is equivalent to \code{getOption("transdoc_to", "en")}.
For \verb{\%to\%}, the non-standard evaluation takes place for \code{to}.}

\item{from}{\code{doc} is translated from the specified language. Currently,
\code{"auto"} is the only supported value and is the default value.}

\item{...}{Arguments passed to the methods.}
}
\description{
The input is translated to a specified language with Google Translate.
For the translation of help document, the source document is derived from
RDocumentation. See \code{language} for the available languages.
}
\examples{
\dontrun{
# Translate a text document
translate("This text will be translated", "ja")
"This text will be translated" \%to\% ja

# Translate a web document
translate("https://transdoc.atusy.net/")
"https://transdoc.atusy.net/" \%to\% ja

# Translate a help document
translate(help(str))
help(str) \%to\% ja
}

}
\references{
Google Translate \url{https://translate.google.com/},
RDocumentation \url{https://www.rdocumentation.org/}
}
